+++ R/cut_the_tree.R
@@ -22,41 +22,12 @@ cut_the_tree <- function(data,
                          delimiter){
 
   # Declare variables used in data.table for R CMD check
-  pathString <- id <- exposed <- NULL
+  exposed <- NULL
 
-  # Extract leafs from pathString
-  tree$leaf <- gsub(paste0(".*(?<=", delimiter, ")(.*)"), "\\1",
-                    tree[["pathString"]],
-                    perl = TRUE)
-
-  # Check that all leafs in data are included on the tree
-  if(any(!(data$leaf %in% tree$leaf))) {
-    cli::cli_abort(
-      c(
-        "x" = "The following leafs are not included on your tree:
-        {unique((data$leaf[!(data$leaf %in% tree$leaf)]))}",
-        "i" = "All leafs must be included in your tree."
-      )
-    )
-  }
-
-  # Combine data and tree
-  temp <- merge(data,
-                tree,
-                by = "leaf",
-                all.x = TRUE)
-
-  # Cut the pathString for each individual and leaf
-  temp[, cut := strsplit(pathString, delimiter, fixed = TRUE)]
-  temp <- temp[, list(cut = unlist(cut)), list(id, exposed)]
-
-  # Include only unique cuts for each individual. This is important to
-  # prevent that events get counted twice when they are on the same path
-  # but on different leafs
-  temp <- unique(temp, by = c("id", "cut"))
+  membership <- get_cut_membership(data, tree, delimiter)
 
   # Calculate number of exposed and unexposed within each cut
-  temp[, list(n0 = sum(exposed == 0), n1 = sum(exposed == 1)),
-       by = "cut"]
+  membership[, list(n0 = sum(exposed == 0), n1 = sum(exposed == 1)),
+             by = "cut"]
 
 }